// src/app/(main)/(collections)/danh-sach-bo-suu-tap/components/CollectionsList/type.ts
/* eslint-disable @typescript-eslint/naming-convention */
export type dataCollections<T> = {
  boardId: string;
  boardName: string;
  createdDateBoard: string;
  boardSlug?:string;
  numberOfProduct: number;
  products: T[];
};

export type Products = {
  bodyStyle: string | null;
  buildDate: string | null;
  contactAddress: string | null;
  contactName: string | null;
  contactPhone: string | null;
  conversionPrice: number | null;
  conversionPriceUnit: string | null;
  countInterested: number | null;
  countViewed: number | null;
  createdDate: string | null;
  createdDateBoardDetail: string | null;
  createdDateProduct: string | null;
  dayId: number | null;
  description: string | null;
  detailLink: string | null;
  districtCode: string | null;
  districtName: string | null;
  domain: string | null;
  doors: string | null;
  drivetrain: string | null;
  engine: string | null;
  expiredDate: string | null;
  exteriorColor: string | null;
  fuelType: string | null;
  imageUrl: string | null;
  interiorColor: string | null;
  isLove: boolean;
  location: string | null;
  make: string | null;
  mileage: string | null;
  model: string | null;
  modifiedDate: string | null;
  modifiedDateProduct: string | null;
  origin: string | null;
  postDate: string | null;
  prdNumber: string | null;
  price: string | null;
  priceUnit: string | null;
  productId: string | null;
  productStatus: string | null;
  productStatusCode: string | null;
  profileLink: string | null;
  profileName: string | null;
  provinceCode: string | null;
  provinceName: string | null;
  seats: string | null;
  site: string | null;
  slug: string | null;
  subDistrictCode: string | null;
  subDistrictName: string | null;
  title: string | null;
  transmission: string | null;
  urlImages: string[] | null;
  useStatus: string | null;
};